9/16/2023 7:53 AM
Added First Visualization Using Seaborn and matplotlib libraries!!!!! Keeping In mind this is my first time creating a project with Python, Simply learning the syntax has helped me apply Python very easily when researching the different functions of these libraries! The visualization also consists of 'number_products' and 'category.' So far, I've created the data cleaning and analysis with one visual! Next, I need to create another visual that demonstrates the most popular products under the categories of both electronics and accessories to see what is top-selling and what we should continue selling more of, but I must ensure that it is neat and organized. Overall analyzing the sales trend over 2019-2020 year!

-George Garrelts